[TNT-224] refactor: 데이터 클래스명 겹치지 않도록 변경
