+++ domain/src/main/java/co/kr/tnt/domain/model/trainee/TraineeDailyRecord.kt
@@ -1,16 +1,16 @@
 package co.kr.tnt.domain.model.trainee
 
-import co.kr.tnt.domain.model.RecordType
+import co.kr.tnt.domain.model.DailyRecord
 import java.time.LocalDate
 import java.time.LocalDateTime
 
 data class TraineeDailyRecord(
     val date: LocalDate,
-    val ptSession: PtSession?,
+    val ptSession: TraineePtSession?,
     val record: List<DailyRecord>,
 )
 
-data class PtSession(
+data class TraineePtSession(
     val ptSessionId: String,
     val trainerName: String,
     val trainerImage: String?,
@@ -18,14 +18,4 @@ data class PtSession(
     val startTime: LocalDateTime,
     val endTime: LocalDateTime,
     val hasRecord: Boolean,
-)
-
-// TODO: RecordType 매핑
-data class DailyRecord(
-    val recordId: String,
-    val recordType: RecordType,
-    val recordTime: LocalDateTime,
-    val recordImage: String?,
-    val recordContents: String,
-    val feedbackCount: Int,
 )
